Refactor DragAndDropHandler
